+++ src/cljs/cxengage_javascript_sdk/domain/topics.cljs
@@ -101,6 +101,7 @@
                  :get-custom-metric-response "cxengage/entities/get-custom-metric-response"
                  :update-custom-metric-response "cxengage/entities/update-custom-metric-response"
                  :get-roles-response "cxengage/entities/get-roles-response"
+                 :get-message-templates-response "cxengage/entities/get-message-templates-response"
                  :get-platform-roles-response "cxengage/entities/get-platform-roles-response"
                  :get-integrations-response "cxengage/entities/get-integrations-response"
                  :get-capacity-rules-response "cxengage/entities/get-capacity-rules-response"